Move delivery method assignment tests to separate file
